App/requirements.txt
pyserial==3.4
django>=2.2.8
django-solo==1.1.3
requests==2.22.0
psycopg2-binary==2.8.4
APScheduler==3.6.3